refactor: separate ui for mask texture editor
