+++ Editor/Inspector/DeleteMeshByMask/DeleteMeshByMaskStyles.uss
@@ -57,6 +57,8 @@
 }
 
 #b-create-mask-texture, #t-edit-mask-texture {
+    margin: 0;
+    padding: 0;
     width: 40px;
 }
