+++ Editor/Inspector/DeleteMeshByMask/MaskTextureEditorOpener.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f189e3acce2314b47b6befaf3c60005e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
