+++ Editor/Inspector/DeleteMeshByMask/MaskTextureEditorOpener.uxml.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 1222e6d40c0e9ee45bffdbaad83cd725
+timeCreated: 1723334567
